Move IS/OS functionality to JavaInputOutputStream
